docker/app/requirements.txt
boto3
psycopg2-binary
pandas
sqlalchemy
awslambdaric 
